+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode/tasks.json
 Animals/Animals/Main
 .vscode/c_cpp_properties.json
+.vscode/settings.json
