+++ Animals/Animals/animalBird.cpp
@@ -11,14 +11,11 @@
 #include <glm/glm.hpp>  // GLM is an optimized math library with syntax to similar to OpenGL Shading Language
 #include <glm/gtc/matrix_transform.hpp> // include this to create transformation matrices
 #include <glm/gtc/type_ptr.hpp>
+
+#include "functions.cpp"
+
 using namespace glm;
 using namespace std;
-
-void SetUniformMat4(GLuint shader_id, const char* uniform_name, mat4 uniform_value);
-void SetUniformVec3(GLuint shader_id, const char* uniform_name, vec3 uniform_value);
-template <class T>
-void SetUniform1Value(GLuint shader_id, const char* uniform_name, T uniform_value);
-
 
 class Bird {
    public:
@@ -188,8 +185,8 @@ class Bird {
          * translate(mat4(1.0f), wingsPositionR) 
          * scale(mat4(1.0f), wingsSize);
          SetUniformMat4(shaderProgram, "worldMatrix", modelMatrix);
-         SetUniformVec3(shaderProgram, "customColor", vec3(0.0f, 1.0f, 1.0f));
-         glUniform1i(texture1Uniform, 1); // Texture unit 2 is now bound to texture1
+         SetUniformVec3(shaderProgram, "customColor", vec3(1.0f, 1.0f, 1.0f));
+         glUniform1i(texture1Uniform, 2); // Texture unit 2 is now bound to texture1
 
          glBindVertexArray(vao);
          glDrawArrays(GL_TRIANGLES, 0, 36); // 36 vertices, starting at index 0
@@ -250,21 +247,3 @@ class Bird {
 
       
 };
-
-// shader variable setters
-void SetUniformMat4(GLuint shader_id, const char* uniform_name, mat4 uniform_value) {
-    glUseProgram(shader_id);
-    glUniformMatrix4fv(glGetUniformLocation(shader_id, uniform_name), 1, GL_FALSE, &uniform_value[0][0]);
-
-}
-void SetUniformVec3(GLuint shader_id, const char* uniform_name, vec3 uniform_value) {
-    glUseProgram(shader_id);
-    glUniform3fv(glGetUniformLocation(shader_id, uniform_name), 1, value_ptr(uniform_value));
-
-}
-template <class T>
-void SetUniform1Value(GLuint shader_id, const char* uniform_name, T uniform_value) {
-  glUseProgram(shader_id);
-  glUniform1i(glGetUniformLocation(shader_id, uniform_name), uniform_value);
-  glUseProgram(0);
-}
